split reviews-model tests into individual files
